+++ app/models/location_photo.rb
@@ -13,4 +13,20 @@ class LocationPhoto < ActiveRecord::Base
 
   # Location Photo Methods
 
+  def number
+    self.location.location_photos.pluck(:id).index(self.id) + 1 rescue -1
+  end
+
+  def next_photo
+    next_index = (self.location.location_photos.pluck(:id).index(self.id) + 1 rescue -1)
+    return nil if next_index < 0
+    self.location.location_photos[next_index]
+  end
+
+  def previous_photo
+    previous_index = (self.location.location_photos.pluck(:id).index(self.id) - 1 rescue -1)
+    return nil if previous_index < 0
+    self.location.location_photos[previous_index]
+  end
+
 end